refactor: move test_mip_template/utils.py content to mip_template/utils.py ...
1. this avoids confusion with two utility scripts

2. sometimes, our users may benefit from read/write data utility funtions (e.g. to create a dat object and call our main function), so they need to be distributed altogether
